+++ src/strongbox/addon.clj
@@ -131,18 +131,40 @@
 
 (defn-spec -load-installed-addons (s/or :ok :addon/toc-list, :error nil?)
   "returns a list of addon data scraped from the .toc files of all addons in given `install-dir`"
-  [install-dir ::sp/addon-dir]
+  [install-dir ::sp/addon-dir, game-track ::sp/game-track]
   (let [addon-dir-list (->> install-dir fs/list-dir (filter fs/directory?) (map str))
         parse-toc (fn [addon-dir]
                     (logging/with-addon {:dirname (-> addon-dir fs/file fs/base-name str)}
-                      (toc/parse-addon-toc-guard addon-dir)))]
-    (->> addon-dir-list (map parse-toc) (remove nil?) vec)))
+                      (let [toc-data-list (toc/parse-addon-toc-guard addon-dir)]
+                        (if (= 1 (count toc-data-list))
+                          ;; whatever toc data we have, we only have one of it (vast majority of cases) so return that
+                          (first toc-data-list)
+
+                          ;; we have multiple sets of toc-data to choose from. which to choose?
+                          ;; prefer the one for the current game track, if it exists, otherwise do as we do in catalogue
+                          ;; and use a list of priorities.
+                          (let [grouped-toc-data (group-by :-toc/game-track toc-data-list)
+                                priority-map {:retail [:retail :classic :classic-tbc]
+                                              :classic [:classic :classic-tbc :retail]
+                                              :classic-tbc [:classic-tbc :classic :retail]}
+                                priorities (get priority-map game-track)
+                                group (utils/first-nn #(get grouped-toc-data %) priorities)]
+                            (when (and (> (count group) 1)
+                                       ;; not all members in group are equal ...
+                                       (not (apply = group)))
+                              (warn (format "multiple sets of different toc data found for %s. using first." game-track)))
+                            (first group))))))]
+    (->> addon-dir-list
+         (map parse-toc)
+         (map #(dissoc % :-toc/game-track))
+         (remove nil?)
+         vec)))
 
 (defn-spec load-installed-addons :addon/toc-list
   "reads the .toc files from the given addon dir, reads any nfo data for 
   these addons, groups them and returns the mooshed data."
-  [install-dir ::sp/extant-dir]
-  (let [addon-list (-load-installed-addons install-dir)
+  [install-dir ::sp/extant-dir, game-track ::sp/game-track]
+  (let [addon-list (-load-installed-addons install-dir game-track)
 
         ;; at this point we have a list of the 'top level' addons, with
         ;; any bundled addons grouped within each one.
@@ -297,9 +319,14 @@
   an 'implicit ignore' is when the addon is under version control or the `.toc` file looks like an unrendered template."
   [install-dir ::sp/extant-dir, addon-dirname ::sp/dirname]
   (let [path (utils/join install-dir addon-dirname)
-        toc-data (toc/parse-addon-toc-guard path)]
-    (or (contains? toc-data :ignore?)
-        (nfo/version-controlled? path))))
+        check-toc-data (fn [[_ filename]]
+                         (-> path
+                             (utils/join filename)
+                             toc/read-toc-file
+                             (toc/parse-addon-toc path)
+                             (contains? :ignore?)))]
+    (or (nfo/version-controlled? path) ;; cheapest check first
+        (->> path toc/find-toc-files (map check-toc-data) utils/any))))
 
 (defn-spec ignore nil?
   "marks the given `addon` and all of it's group members (if any) as 'ignored'"
@@ -405,7 +432,8 @@
   "returns `true` when given `addon` can be updated to a newer version."
   [addon map?] ;; deliberately lenient
   (let [{:keys [installed-version pinned-version version
-                game-track installed-game-track]} addon]
+                game-track installed-game-track
+                supported-game-tracks]} addon]
     (cond
       ;; not expanded
       (not version) false
@@ -418,8 +446,14 @@
       pinned-version (and (not= pinned-version installed-version)
                           (= pinned-version version))
 
-      (and game-track installed-game-track) (not= [version game-track]
-                                                  [installed-version installed-game-track])
+      ;; if the available version is equal to the installed version
+      ;; AND the installed game track is the same as the available game track
+      ;; (OR, the installed game track is present in the list of game tracks supported by the addon),
+      ;; then the addon is NOT updateable.
+      (and game-track installed-game-track)
+      (not (and (= version installed-version)
+                (or (= game-track installed-game-track)
+                    (utils/in? installed-game-track supported-game-tracks))))
 
       :else (not= version installed-version))))