+++ test/strongbox/addon_test.clj
@@ -4,6 +4,7 @@
    ;;[taoensso.timbre :as log :refer [debug info warn error spy]]
    [me.raynes.fs :as fs]
    [strongbox
+    [zip :as zip]
     [logging :as logging]
     [utils :as utils]
     [test-helper :as helper :refer [fixture-path slurp-fixture helper-data-dir with-running-app]]
@@ -42,66 +43,86 @@
 
 (deftest group-addons
   (testing "addons with nothing to group on are not modified"
-    (let [addon-list [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"}
-                      {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"}
-                      {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"}]
+    (let [addon-list [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3" :supported-game-tracks [:retail]}
+                      {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6" :supported-game-tracks [:retail]}
+                      {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9" :supported-game-tracks [:retail]}]
           expected addon-list]
       (is (= expected (addon/group-addons addon-list)))))
 
   (testing "addons with groupable data but no groupings are not modified"
     (let [addon-list [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? true}
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "bar" :primary? true}
                       {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"
+                       :supported-game-tracks [:retail]
                        :group-id "baz" :primary? true}]
           expected addon-list]
       (is (= expected (addon/group-addons addon-list)))))
 
   (testing "addons with groupable data with one marked as the `primary`, group as expected"
     (let [addon-list [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? true}
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}
                       {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"
+                       :supported-game-tracks [:retail]
                        :group-id "bar" :primary? true}]
 
           expected [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                     :supported-game-tracks [:retail]
                      :group-id "foo" :primary? true :group-addon-count 2 :group-addons
                      [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? true}
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}]}
 
                     {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"
+                     :supported-game-tracks [:retail]
                      :group-id "bar" :primary? true}]]
       (is (= expected (addon/group-addons addon-list)))))
 
   (testing "synthetic records are created for groupable addons with no primary addon"
     (let [addon-list [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}
                       {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"
+                       :supported-game-tracks [:retail]
                        :group-id "bar" :primary? true}]
 
           expected [{:name "a1", :dirname "A1", :label "foo (group)", :description "group record for the foo addon" :interface-version 80300 :installed-version "1.2.3"
+                     :supported-game-tracks [:retail]
                      :group-id "foo" :primary? false :group-addon-count 2 :group-addons
                      [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}]}
 
                     {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"
+                     :supported-game-tracks [:retail]
                      :group-id "bar" :primary? true}]]
       (is (= expected (addon/group-addons addon-list)))))
 
   (testing "if any one addon in a group is ignored, the top-level addon ('all') is also ignored"
     (let [addon-list [{:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? true}
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false}
                       {:name "a3", :dirname "A3", :label "A2", :description "" :interface-version 80300 :installed-version "7.8.9"
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? false :ignore? true}]
 
           expected [{:name "a1"
@@ -110,6 +131,8 @@
                      :description ""
                      :interface-version 80300
                      :installed-version "1.2.3"
+
+                     :supported-game-tracks [:retail]
                      :group-id "foo"
                      :primary? true
                      :ignore? true
@@ -119,6 +142,8 @@
                                      :label "A1",
                                      :description ""
                                      :interface-version 80300
+
+                                     :supported-game-tracks [:retail]
                                      :installed-version "1.2.3"
                                      :group-id "foo"
                                      :primary? true}
@@ -127,6 +152,8 @@
                                      :label "A2",
                                      :description ""
                                      :interface-version 80300
+
+                                     :supported-game-tracks [:retail]
                                      :installed-version "4.5.6"
                                      :group-id "foo"
                                      :primary? false}
@@ -135,6 +162,8 @@
                                      :label "A2",
                                      :description ""
                                      :interface-version 80300
+
+                                     :supported-game-tracks [:retail]
                                      :installed-version "7.8.9"
                                      :group-id "foo"
                                      :primary? false
@@ -152,8 +181,15 @@
           some-addon-toc (utils/join some-addon-path "SomeAddon.toc")
           _ (spit some-addon-toc "## Title: SomeAddon\n## Description: asdf\n## Interface: 80300\n## Version: 1.2.3")
 
-          expected [{:name "someaddon", :dirname "SomeAddon", :label "SomeAddon", :description "asdf", :interface-version 80300, :installed-version "1.2.3"}]]
-      (is (= expected (addon/load-installed-addons addon-dir))))))
+          expected [{:name "someaddon",
+                     :dirname "SomeAddon",
+                     :label "SomeAddon",
+                     :description "asdf",
+                     :interface-version 80300,
+
+                     :supported-game-tracks [:retail]
+                     :installed-version "1.2.3"}]]
+      (is (= expected (addon/load-installed-addons addon-dir :retail))))))
 
 (deftest load-installed-addons-2
   (testing "toc data and nfo data are mooshed together as expected"
@@ -175,13 +211,16 @@
           _ (spit some-addon-nfo (utils/to-json nfo-data))
 
           expected [{;; toc data
-                     :name "someaddon",
-                     :dirname "SomeAddon",
-                     :label "SomeAddon",
-                     :description "asdf",
-                     :interface-version 80300,
+                     :name "someaddon"
+                     :dirname "SomeAddon"
+                     :label "SomeAddon"
+                     :description "asdf"
+                     :interface-version 80300
+
 
                      ;; shared between toc and nfo, nfo wins out
+
+
                      :installed-version "1.2.3"
 
                      ;; unique items from nfo data
@@ -189,8 +228,9 @@
                      :source-id 123
                      :group-id "fdsa"
                      :installed-game-track :retail
+                     :supported-game-tracks [:retail]
                      :primary? true}]]
-      (is (= expected (addon/load-installed-addons addon-dir))))))
+      (is (= expected (addon/load-installed-addons addon-dir :retail))))))
 
 (deftest load-installed-addons--invalid-nfo-data-not-loaded
   (testing "invalid nfo data is not loaded"
@@ -213,8 +253,9 @@
                     }
           _ (spit some-addon-nfo (utils/to-json nfo-data))
 
-          expected [{:name "someaddon", :dirname "SomeAddon", :label "SomeAddon", :description "asdf", :interface-version 80300, :installed-version "1.2.3"}]]
-      (is (= expected (addon/load-installed-addons addon-dir))))))
+          expected [{:name "someaddon", :dirname "SomeAddon", :label "SomeAddon", :description "asdf", :interface-version 80300, :installed-version "1.2.3"
+                     :supported-game-tracks [:retail]}]]
+      (is (= expected (addon/load-installed-addons addon-dir :retail))))))
 
 (deftest load-installed-addons--explicit-nfo-ignore
   (testing "ignore flag in nfo data overrides any ignore flag in toc data"
@@ -231,10 +272,30 @@
                                                  :ignore? false})) ;; expressly un-ignoring this otherwise-ignored addon
 
           expected [{:name "someaddon", :dirname "SomeAddon", :label "SomeAddon", :description "asdf", :interface-version 80300
+                     :supported-game-tracks [:retail]
                      :installed-version "@project-version@"
                      :source "curseforge" :source-id 123
                      :ignore? false}]]
-      (is (= expected (addon/load-installed-addons addon-dir))))))
+      (is (= expected (addon/load-installed-addons addon-dir :retail))))))
+
+(deftest load-installed-addons--multiple-non-identical-toc-data
+  (let [fixture (helper/fixture-path "everyaddon--1-2-3--multi-toc--inconsistent.zip")
+        game-track :classic
+        expected [;; description has been modified in "-Classic" vs "-Vanilla"
+                  {:description "Slightly differently does what no other addon does."
+                   :dirname "EveryAddon",
+                   :installed-version "1.2.3",
+                   :interface-version 11307,
+                   :label "EveryAddon 1.2.3",
+                   :name "everyaddon",
+                   :supported-game-tracks [:classic :classic-tbc :retail]}]
+
+        expected-warning "multiple sets of different toc data found for :classic. using first."]
+    (zip/unzip-file fixture (helper/install-dir))
+    (let [[warning] (logging/buffered-log
+                     :warn
+                     (is (= expected (addon/-load-installed-addons (helper/install-dir) game-track))))]
+      (is (= expected-warning warning)))))
 
 (deftest remove-addon--malign-addon-data
   (testing "uninstalling an addon whose `:dirname` value is corrupted (somehow) shouldn't affect data outside of the addon dir"
@@ -254,7 +315,8 @@
 
           expected (helper/install-dir-contents)
 
-          defaults {:name "nom" :label "Nom" :description "" :interface-version 90100 :installed-version "0.1"}]
+          defaults {:name "nom" :label "Nom" :description "" :interface-version 90100 :installed-version "0.1"
+                    :supported-game-tracks [:retail]}]
       (doseq [[given error-prefix] cases]
         (let [[error-message]
               (logging/buffered-log :error
@@ -267,21 +329,26 @@
 (deftest test-pinned-dir-list
   (testing "directory names of pinned addons are detected"
     (let [addon-list [;; single, unpinned, addon
-                      {:name "a1", :dirname "A1", :label "A1", :description "" :interface-version 80300 :installed-version "1.2.3"
+                      {:name "a1", :dirname "A1", :label "A1", :description "" :installed-version "1.2.3" :interface-version 80300
+                       :supported-game-tracks [:retail]
                        :group-id "foo" :primary? true}
 
                       ;; single, pinned, addon
                       {:name "a2", :dirname "A2", :label "A2", :description "" :interface-version 80300 :installed-version "4.5.6"
+                       :supported-game-tracks [:retail]
                        :group-id "bar" :primary? false :pinned-version "4.5.6"}
 
                       ;; grouped addon, group members pinned
                       {:name "a3", :dirname "A3", :label "A3", :description "" :interface-version 80300 :installed-version "7.8.9"
+                       :supported-game-tracks [:retail]
                        :group-id "baz" :primary? true, :pinned-version "7.8.9"
                        :group-addons [;; addon's contain themselves in `:group-addons`
                                       {:name "a3", :dirname "A3", :label "A3", :description "" :interface-version 80300 :installed-version "7.8.9"
+                                       :supported-game-tracks [:retail]
                                        :group-id "baz" :primary? true :pinned-version "7.8.9"}
 
                                       {:name "a3-sub", :dirname "A3_Sub", :label "A3-Sub", :description "" :interface-version 80300 :installed-version "7.8.9.0"
+                                       :supported-game-tracks [:retail]
                                        :group-id "baz" :primary? false :pinned-version "7.8.9"}]}
 
                       ;; abnormal case.
@@ -289,12 +356,15 @@
                       ;; todo: support this case and consider both A4 and A4_Sub pinned?
                       ;; is this case possible if an unpinned addon overwrites a pinned one?
                       {:name "a4", :dirname "A4", :label "A4", :description "" :interface-version 80300 :installed-version "0.1.2"
+                       :supported-game-tracks [:retail]
                        :group-id "bup" :primary? false
                        :group-addons [;; addon's contain themselves in `:group-addons`
                                       {:name "a4", :dirname "A4", :label "A4", :description "" :interface-version 80300 :installed-version "0.1.2"
+                                       :supported-game-tracks [:retail]
                                        :group-id "bup" :primary? false} ;;, :pinned-version "7.8.9"} ;; we're not doing this. should we be doing this?
 
                                       {:name "a4-sub", :dirname "A4_Sub", :label "A4-Sub", :description "" :interface-version 80300 :installed-version "0.1.2.0"
+                                       :supported-game-tracks [:retail]
                                        :group-id "bup" :primary? false :pinned-version "foooooooooooo"}]}]
 
           expected #{"A2" "A3" "A3_Sub"
@@ -311,8 +381,15 @@
 (deftest test-overwrites-pinned?
   (testing "addon zip files that would extract over a pinned addon are correctly detected"
     (let [downloaded-file (fixture-path "everyaddon--1-2-3.zip")
-          addon {:name "EveryAddon", :dirname "EveryAddon", :label "Every Addon", :description ""
-                 :interface-version 80300 :installed-version "1.2.3" :group-id "foo" :primary? true}
+          addon {:name "EveryAddon",
+                 :dirname "EveryAddon",
+                 :label "Every Addon",
+                 :description ""
+                 :interface-version 80300
+                 :installed-version "1.2.3"
+                 :supported-game-tracks [:retail]
+                 :group-id "foo"
+                 :primary? true}
           pinned-addon (assoc addon :pinned-version "1.2.3")]
       (is (addon/overwrites-pinned? downloaded-file [pinned-addon]))
       (is (not (addon/overwrites-pinned? downloaded-file [addon]))))))
@@ -435,3 +512,24 @@
   (is (addon/unpinnable? {:pinned-version "1.2.3"}))
   (is (not (addon/unpinnable? {:pinned-version "1.2.3" :ignore? true})))
   (is (not (addon/unpinnable? {}))))
+
+(deftest implicitly-ignored?
+  (testing "evidence of a template in the toc file marks addon as implicitly ignored"
+    (let [nom "EveryAddon"
+          addon-dir (utils/join (helper/install-dir) nom)
+          path (fn [bit]
+                 (utils/join addon-dir bit))]
+      (fs/mkdir addon-dir)
+      (spit (path "EveryAddon.toc") "## Title: Foo\n## Version: @project-version@")
+      (is (true? (addon/implicitly-ignored? (helper/install-dir) nom))))))
+
+(deftest implicitly-ignored?--multi-toc
+  (testing "evidence of a template in *any* toc file, marks addon as implicitly ignored"
+    (let [nom "EveryAddon"
+          addon-dir (utils/join (helper/install-dir) nom)
+          path (fn [bit]
+                 (utils/join addon-dir bit))]
+      (fs/mkdir addon-dir)
+      (spit (path "EveryAddon.toc") "## Title: Foo\n## Version: 1.2.3")
+      (spit (path "EveryAddon-Mainline.toc") "## Title: Foo-Bar\n## Version: @project-version@")
+      (is (true? (addon/implicitly-ignored? (helper/install-dir) nom))))))